+++ CHANGELOG.md
@@ -115,6 +115,8 @@
 * `[jest-environment-jsdom]` Update JSOM to 11.4, which includes built-in
   support for `requestAnimationFrame`
   ([#4919](https://github.com/facebook/jest/pull/4919))
+* `[jest-snapshot]` Promises support for `toThrowErrorMatchingSnapshot`
+  ([#4946](https://github.com/facebook/jest/pull/4946))
 
 ### Chore & Maintenance